Dont append full wrapped blobtx to filters (#9471)
In txpool, the rlp is appended to be broadcasted as is. This PR removes
the unnecessary parts of the wrapped rlp, namely everything but the
payload part.

There is a redundancy in rlp libraries inside and outside `erigon-lib`
directory. I have added some redundant code here, for future reference
within `core/types`