e3: adding the proper pruned node (known as full node) (#10165)
This PR adds the following feauture:

## Limited downloading of snapshot files

If the user requests a pruned node, we avoid downloading the snapshot
files that are below the range requested. We do this by building a
`blacklist` and then just skip all snapshots in such blacklist.

#### IMPORTANT

due to the different unit of measurement (block number vs step size), we
need to first download headers and bodies, and then convert block number
of blocks to keep to steps by iterating over the bodies snapshots
previously downloaded.

### Pruning of block snapshots

Now blocks can be pruned through the `Delete` method added to the
`Snapshots` struct. this happens as part of block retirement at the end
after `retireBlocks` so that we do not cause nasty race conditions. it
is injected through a function reference in the `retire` flow.