Improved `eth_syncing` (#10398)
The approach is simple, every time the `Execution` module sees a block
it saves the number in DB. That number is the heuristics for the "tip of
the chain".

Also changes from stages.Finish to stages.Execution. the issue was that
for syncing up everything (stages.Headers as well) is in a single mdbx
tx. Now we have separate key which is committed in a small `Update`
before the processing.