E3: Remotion of KeysTable in `Domain` (#10771)
# PR: Remove `keysTable` from Domain

This PR removes the use of `keysTable` from Domain. Below is how this
was achieved.

## Modifications to `ValsTable`

Now, `ValsTable` in domains are dupsorted and have the following layout:

* `Key -> Inverted_Step + Value`

This allows us to maintain the same ordering with respect to keys while
also keeping multiple versions of each key. This slightly improves cases
where one key has multiple values over different steps, as we decouple
the key. Most importantly, it allows us to track each version of the key
without needing another table, thereby eliminating the external step.

### `ValsTable`: Code (largeVals=true)

For `Code`, this is not possible because it has large values, which
makes MDBX problematic. Therefore, the `Code` table is kept as
non-dupsort but modified with the layout:

* `Key + Inverted_Step -> Value`

However, this only works if keys are of fixed size. This is because, in
lexicographical order, `Seek` prioritizes lexicography over length. For
example:

* `AAAB + FFFFFA` (Key + Inv_Step)
* `AAABA + FFFFFA`

If we seek `AAAB`, we will actually get the value for `AAABA`, as the
inverted step interferes.

### Collate

In collate, for largeVals, we need to sort the keys afterward (for the
same reason). Maybe we can replace that with a collector to avoid the
extra linear RAM usage for `code`.

## Results

* One less random read during `getLatestFromDB`.
* ChainData size after 10 days of running (2 prunes happened):
```
admin@erigon5900d:~/erigon$ du  gg/chaindata/  -ch
12G     gg/chaindata/
12G     total
```
* Faster flushing of domains by about 25%

---------

Co-authored-by: alex.sharov <AskAlexSharov@gmail.com>